@@ -10,7 +10,7 @@ authors = [
 version = "26"
 description = "Build Bespoke OS Images"
 readme = "README.md"
-requires-python = ">=3.9"
+requires-python = ">=3.10"
 license = {text = "LGPL-2.1-or-later"}
 
 [project.optional-dependencies]
@@ -53,7 +53,7 @@ multi_line_output = 3
 py_version = "39"
 
 [tool.pyright]
-pythonVersion = "3.9"
+pythonVersion = "3.10"
 include = [
     "mkosi/**/*.py",
     "tests/**/*.py",
@@ -61,7 +61,7 @@ include = [
 ]
 
 [tool.mypy]
-python_version = 3.9
+python_version = "3.10"
 # belonging to --strict
 warn_unused_configs = true
 disallow_any_generics = true
