@@ -10,7 +10,6 @@ import tempfile
 import textwrap
 from collections.abc import Iterator, Mapping, Sequence
 from pathlib import Path
-from typing import Optional
 
 from mkosi.config import (
     BiosBootloader,
@@ -168,7 +167,7 @@ def want_grub_bios(context: Context, partitions: Sequence[Partition] = ()) -> bo
     return (have and bios and esp and root and installed) if partitions else have
 
 
-def find_grub_directory(context: Context, *, target: str) -> Optional[Path]:
+def find_grub_directory(context: Context, *, target: str) -> Path | None:
     for d in ("usr/lib/grub", "usr/share/grub2"):
         if (p := context.root / d / target).exists() and any(p.iterdir()):
             return p
@@ -176,7 +175,7 @@ def find_grub_directory(context: Context, *, target: str) -> Optional[Path]:
     return None
 
 
-def find_grub_binary(config: Config, binary: str) -> Optional[Path]:
+def find_grub_binary(config: Config, binary: str) -> Path | None:
     assert "grub" not in binary
 
     # Debian has a bespoke setup where if only grub-pc-bin is installed, grub-bios-setup is installed in
@@ -185,7 +184,7 @@ def find_grub_binary(config: Config, binary: str) -> Optional[Path]:
     return config.find_binary(f"grub-{binary}", f"grub2-{binary}", f"/usr/lib/grub/i386-pc/grub-{binary}")
 
 
-def prepare_grub_config(context: Context) -> Optional[Path]:
+def prepare_grub_config(context: Context) -> Path | None:
     config = context.root / "efi" / context.config.distribution.installer.grub_prefix() / "grub.cfg"
     with umask(~0o700):
         config.parent.mkdir(exist_ok=True)
@@ -223,8 +222,8 @@ def grub_mkimage(
     *,
     target: str,
     modules: Sequence[str] = (),
-    output: Optional[Path] = None,
-    sbat: Optional[Path] = None,
+    output: Path | None = None,
+    sbat: Path | None = None,
 ) -> None:
     mkimage = find_grub_binary(context.config, "mkimage")
     assert mkimage
@@ -293,7 +292,7 @@ def grub_mkimage(
         )  # fmt: skip
 
 
-def find_signed_grub_image(context: Context) -> Optional[Path]:
+def find_signed_grub_image(context: Context) -> Path | None:
     arch = context.config.architecture.to_efi()
 
     patterns = [
@@ -476,9 +475,9 @@ def run_systemd_sign_tool(
     *,
     cmdline: Sequence[PathString],
     options: Sequence[PathString],
-    certificate: Optional[Path],
+    certificate: Path | None,
     certificate_source: CertificateSource,
-    key: Optional[Path],
+    key: Path | None,
     key_source: KeySource,
     env: Mapping[str, str] = {},
     stdout: _FILE = None,
