@@ -3,7 +3,6 @@
 import textwrap
 from collections.abc import Sequence
 from pathlib import Path
-from typing import Optional
 
 from mkosi.config import Cacheonly, Config
 from mkosi.context import Context
@@ -63,7 +62,7 @@ class Dnf(PackageManager):
         context: Context,
         repositories: Sequence[RpmRepository],
         filelists: bool = True,
-        metadata_expire: Optional[str] = None,
+        metadata_expire: str | None = None,
     ) -> None:
         (context.sandbox_tree / "etc/dnf/vars").mkdir(parents=True, exist_ok=True)
         (context.sandbox_tree / "etc/yum.repos.d").mkdir(parents=True, exist_ok=True)
