@@ -4,7 +4,6 @@ import os
 import textwrap
 import uuid
 from pathlib import Path
-from typing import Optional
 
 import pytest
 
@@ -49,7 +48,7 @@ from mkosi.distribution import Distribution
 
 
 @pytest.mark.parametrize("path", [None, "/baz/qux"])
-def test_args(path: Optional[Path]) -> None:
+def test_args(path: Path | None) -> None:
     dump = textwrap.dedent(
         f"""\
         {{
